rename files and paths for shared/message
