+++ feature-libs/organization/administration/components/cost-center/details/cost-center-details.component.spec.ts
@@ -7,8 +7,8 @@ import { of, Subject } from 'rxjs';
 import { CardTestingModule } from '../../shared/card/card.testing.module';
 import { ToggleStatusModule } from '../../shared/detail/toggle-status-action/toggle-status.module';
 import { OrganizationItemService } from '../../shared/organization-item.service';
-import { MessageTestingModule } from '../../shared/organization-message/message.testing.module';
-import { MessageService } from '../../shared/organization-message/services/message.service';
+import { MessageTestingModule } from '../../shared/message/message.testing.module';
+import { MessageService } from '../../shared/message/services/message.service';
 import { CostCenterDetailsComponent } from './cost-center-details.component';
 import createSpy = jasmine.createSpy;
