+++ feature-libs/organization/administration/components/shared/card/card.component.spec.ts
@@ -8,7 +8,7 @@ import { IconTestingModule } from 'projects/storefrontlib/src/cms-components/mis
 import { ViewComponent } from 'projects/storefrontlib/src/shared/components/split-view/view/view.component';
 import { of } from 'rxjs';
 import { OrganizationItemService } from '../organization-item.service';
-import { MessageTestingModule } from '../organization-message/message.testing.module';
+import { MessageTestingModule } from '../message/message.testing.module';
 import { CardComponent } from './card.component';
 import createSpy = jasmine.createSpy;
