+++ feature-libs/organization/administration/components/shared/card/card.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { I18nModule, UrlModule } from '@spartacus/core';
 import { IconModule, SplitViewModule } from '@spartacus/storefront';
-import { MessageModule } from '../organization-message/message.module';
+import { MessageModule } from '../message/message.module';
 import { CardComponent } from './card.component';
 
 /**
