+++ feature-libs/organization/administration/components/shared/detail/toggle-status-action/index.ts
@@ -1,4 +1,4 @@
-export * from '../../organization-message/confirmation/confirmation-message.component';
-export * from '../../organization-message/confirmation/confirmation-message.module';
+export * from '../../message/confirmation/confirmation-message.component';
+export * from '../../message/confirmation/confirmation-message.module';
 export * from './toggle-status.component';
 export * from './toggle-status.module';
