+++ feature-libs/organization/administration/components/shared/detail/toggle-status-action/toggle-status.module.ts
@@ -2,8 +2,8 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { I18nModule, UrlModule } from '@spartacus/core';
-import { ConfirmationMessageModule } from '../../organization-message/confirmation/confirmation-message.module';
-import { MessageModule } from '../../organization-message/message.module';
+import { ConfirmationMessageModule } from '../../message/confirmation/confirmation-message.module';
+import { MessageModule } from '../../message/message.module';
 import { ToggleStatusComponent } from './toggle-status.component';
 
 @NgModule({
