+++ feature-libs/organization/administration/components/shared/index.ts
@@ -6,7 +6,7 @@ export * from './detail/index';
 export * from './form/index';
 export * from './organization-item.service';
 export * from './list/index';
-export * from './organization-message/index';
+export * from './message/index';
 export * from './organization-sub-list/index';
 export * from './organization-table/index';
 export * from './organization.model';
