+++ feature-libs/organization/administration/components/shared/item-exists.directive.spec.ts
@@ -5,7 +5,7 @@ import { of } from 'rxjs/internal/observable/of';
 import { Subject } from 'rxjs/internal/Subject';
 import { ItemExistsDirective } from './item-exists.directive';
 import { OrganizationItemService } from './organization-item.service';
-import { MessageService } from './organization-message/services/message.service';
+import { MessageService } from './message/services/message.service';
 import { GlobalMessageType } from '@spartacus/core';
 import createSpy = jasmine.createSpy;
