+++ feature-libs/organization/administration/components/shared/organization-sub-list/organization-sub-list.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { I18nModule } from '@spartacus/core';
 import { PaginationModule, TableModule } from '@spartacus/storefront';
 import { CardModule } from '../card/card.module';
-import { MessageModule } from '../organization-message/message.module';
+import { MessageModule } from '../message/message.module';
 import { AssignCellComponent } from './assign-cell.component';
 import { OrganizationSubListComponent } from './organization-sub-list.component';
