+++ feature-libs/organization/administration/components/user/details/user-details.component.spec.ts
@@ -8,8 +8,8 @@ import { of, Subject } from 'rxjs';
 import { CardTestingModule } from '../../shared/card/card.testing.module';
 import { ToggleStatusModule } from '../../shared/detail/toggle-status-action/toggle-status.module';
 import { OrganizationItemService } from '../../shared/organization-item.service';
-import { MessageTestingModule } from '../../shared/organization-message/message.testing.module';
-import { MessageService } from '../../shared/organization-message/services/message.service';
+import { MessageTestingModule } from '../../shared/message/message.testing.module';
+import { MessageService } from '../../shared/message/services/message.service';
 import { ItemExistsDirective } from '../../shared/item-exists.directive';
 import { UserDetailsComponent } from './user-details.component';
 import createSpy = jasmine.createSpy;
